adds template/tasks endpoint
