+++ api/server.js
@@ -3,7 +3,7 @@ const cors = require('cors')
 const helmet = require('helmet')
 
 const projectsRouter = require('../routes/projects-routes')
-const taskRouter = require('../routes/tasks-routes')
+const templateRouter = require('../routes/template-router')
 const register = require('../auth/authRouter')
 const login = require('../auth/authRouter')
 const weather = require('../routes/weather')
@@ -26,7 +26,7 @@ server.use('/api',login,register)
 
 // Add auth to routes
 server.use('/api/auth',projectsRouter)
-server.use('/api/auth',taskRouter)
+server.use('/api/auth',templateRouter)
 server.use('/api/auth/:uid',weather)
