+++ routes/projects-routes.js
@@ -62,10 +62,7 @@ router.post('/:uid/projects',  async (req,res)=>{
        let projectID = req.body.projectID
        let body = req.body
        let uid = req.params.uid
-       let tasks = [
-             {"due_date":String,"task_description":String,"task_name":String}        
-    
-    ]
+
   await  dbRef.child(`/${uid}/projects/${projectID}`).set(
                     
                       {
